% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO__functions.r
\name{cdo.dates}
\alias{cdo.dates}
\title{Extract dates from a file using CDO}
\usage{
cdo.dates(f)
}
\arguments{
\item{f}{Filename}
}
\description{
Extracts the timestaps from a file using CDO and coverts them to a Date format suitable for use
elsewhere in R. It can be advantageous to use this approach, rather than, e.g. raster, when using CDO
as your main processing tool, to ensure consistency in the approach. This is also rather handy when
dealing with non-standard calendars that raster doesn't support.
}
