% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO__functions.r
\docType{methods}
\name{griddes}
\alias{griddes}
\alias{griddes,Raster-method}
\alias{griddes,Extent-method}
\title{Create a CDO compatable grid descriptor}
\usage{
griddes(x, ...)

\S4method{griddes}{Raster}(x, ...)

\S4method{griddes}{Extent}(x, res = 1, ...)
}
\arguments{
\item{x}{Provides information to define the grid}

\item{res}{The resolution of the grid (single number of vector of two numbers)}
}
\description{
Creates a grid descriptor that can be used by CDO in regridding operations
}
