% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdump.r
\name{ncdump}
\alias{ncdump}
\alias{ncdump.coordinates}
\alias{ncdump.times}
\title{Extract metadata using ncdump}
\usage{
ncdump.coordinates(f)

ncdump.times(f)

ncdump(f)
}
\arguments{
\item{f}{Path to the netcdf file to process}
}
\value{
`ncdump.coordinates` returns a named list of the coordinates. `ncdump.time` returns just the
time dimension, if it exists
}
\description{
ncdump is a utility provided when installing the NetCDF libraries that is very handy
for extracting metadata information from a netcdf file
}
