% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCO_functions.r
\name{nco}
\alias{nco}
\alias{set.nco.defaults}
\alias{ncks}
\alias{ncwa}
\alias{ncra}
\alias{ncrcat}
\alias{ncecat}
\alias{nces}
\alias{ncbo}
\alias{ncdiff}
\alias{ncap2}
\alias{ncrename}
\alias{ncatted}
\title{NCO operators}
\usage{
set.nco.defaults(defaults = "")

ncks(..., debug = FALSE)

ncwa(..., debug = FALSE)

ncra(..., debug = FALSE)

ncrcat(..., debug = FALSE)

ncecat(..., debug = FALSE)

nces(..., debug = FALSE)

ncbo(..., debug = FALSE)

ncdiff(..., debug = FALSE)

ncap2(..., debug = FALSE)

ncrename(..., debug = FALSE)

ncatted(..., debug = FALSE)
}
\arguments{
\item{defaults}{Character string supplying the defaults to be used by CDO}

\item{...}{List of arguments to be converted to a text string}

\item{debug}{Doesn't run the command, but returns the command that would be run}
}
\description{
Provides an interface to the (external to R) NetCDF Operators (NCO)
}
\details{
Default options for nco can be set at the system wide level
using the \code{set.nc.defaults()} command. Two of the most useful
of these include "--overwrite" (to overwrite files by default) and
"--netcdf4" to force NetCDF4 output. See the documentation of NCO for more.
}
\examples{
# Note, to run this in the wild, remove the debug=TRUE arguments
set.nco.defaults("--overwrite")
in.fname <- "tos_Omon_IPSL-CM5A-LR_historical_r1i1p1_185001-200512.nc"
out.fname <- "foo.nc"
ncra(in.fname,out.fname,debug=TRUE)
ncwa(in.fname,out.fname,debug=TRUE)
}
