% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO__functions.r
\name{cdo}
\alias{cdo}
\alias{set.cdo.defaults}
\alias{csl}
\alias{ssl}
\title{Climate Data Operators (CDO)}
\usage{
cdo(..., debug = FALSE)

set.cdo.defaults(defaults = "")

csl(...)

ssl(...)
}
\arguments{
\item{debug}{Doesn't run the command, but returns the command that would be run.}

\item{defaults}{Character string supplying the defaults to be used by CDO}
}
\description{
Provides an interface for running commands suitable for use with
the (externally installed) Climate Data Operators (CDO)
}
\details{
Default options for cdo can be set at the system wide level
using the \code{set.cdo.defaults()} command. Two of the most useful
of these include "-O" (to overwrite files by default) and "-s" to
silence the outputs of cdo. See the helpfiles of CDO for more.

\code{csl} converts a set of arguments into a comma separated list

\code{ssl} converts a set of arguments into a space separated list
}
\examples{
# What version of CDO am I running?
cdo("--version")

# Select a variable from a CMIP5 model
# Note, to run this in the wild, remove the debug=TRUE argument
in.fname <- "tos_Omon_IPSL-CM5A-LR_historical_r1i1p1_185001-200512.nc"
out.fname <- "foo.nc"
cdo("selvar",in.fname,out.fname,debug=TRUE)

# Use the defaults and csl to build up a more complex argument
# Note, to run this in the wild, remove the debug=TRUE argument
in.fname <- "tos_Omon_IPSL-CM5A-LR_historical_r1i1p1_185001-200512.nc"
out.fname <- "foo.nc"
set.cdo.defaults("--silent")
yrs <- seq(1950,2005,by=5) #Every fifth year
cdo(csl("selyear",yrs),in.fname,out.fname,debug=TRUE)
}
